[20631] US #998: USH-OAI test af solr høstning - Gui til Solr Test høstning
git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@104831 b87a6194-a221-0410-ad6f-3b8ec35bba13

+++ commons/utils/ush-solr-harvester-service-connector/src/main/java/dk/dbc/dataio/ushsolrharvester/service/connector/UshSolrHarvesterServiceConnector.java
@@ -63,12 +63,12 @@ public class UshSolrHarvesterServiceConnector {
 
     /**
      * Runs a Solr test harvest
-     * @param id the is of the harvest to test run
+     * @param id the id of the harvest to test run
      * @return the harvester id
      * @throws ProcessingException on general communication error
      * @throws UshSolrHarvesterServiceConnectorException on failure
      */
-    public String runTestHarvest(final int id) throws ProcessingException, UshSolrHarvesterServiceConnectorException {
+    public String runTestHarvest(final long id) throws ProcessingException, UshSolrHarvesterServiceConnectorException {
         final String NO_DATA = "";
         log.trace("UshSolrHarvesterServiceConnector: runTestHarvest({});", id);
         final StopWatch stopWatch = new StopWatch();
